afxdp/bpf_map: relocate inline tests to bpf_map_tests.rs
Same playbook. 206-LOC inline test block extracted.
After: bpf_map.rs drops from 1,389 → 1,186 LOC.
cargo build/test 92/878.
